+++ packages/notform/CHANGELOG.md
@@ -1,5 +1,12 @@
 # notform
 
+## 1.0.0-alpha.8
+
+### Patch Changes
+
+- [#8](https://github.com/favorodera/notform/pull/8) [`3f74afa`](https://github.com/favorodera/notform/commit/3f74afa91218db1ecc59beb10fdd63430185e5d1) Thanks [@favorodera](https://github.com/favorodera)! - - Refactored `NotForm` props to accept `asChild` and `as` options to to implement dynamic rendering
+  - Created a standard schema based simple validator for testing and removed all "library inclined" tests
+
 ## 1.0.0-alpha.7
 
 ### Patch Changes